Improve: add useAdminProducts composable
